requirements.txt
boto3==1.17.8
botocore==1.20.8
certifi==2020.12.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
django-extensions==3.1.1
django-heroku==0.3.1
django-storages==1.11.1
django-taggit==1.3.0
django-toolbelt==0.0.1
docutils==0.16
gunicorn==20.0.4
jmespath==0.10.0
pbr==5.5.1
Pillow==8.1.0
pipenv==2020.11.15
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
static3==0.7.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
